# [FasterTransformer] Update FasterTransformer and add performance info logger