--- a/examples/machine_translation/transformer/faster_transformer/encoder_decoding_predict.py
+++ b/examples/machine_translation/transformer/faster_transformer/encoder_decoding_predict.py
@@ -15,6 +15,7 @@ from pprint import pprint
 from paddlenlp.transformers import TransformerModel
 from paddlenlp.transformers import position_encoding_init
 from paddlenlp.ops import FasterTransformer
+from paddlenlp.utils.log import logger
 
 sys.path.append("../")
 import reader
@@ -55,6 +56,12 @@ def parse_args():
         type=float,
         help="The probability threshold for topp_sampling. Default is 0.0 which means it won't go through topp_sampling. "
     )
+    parser.add_argument(
+        "--batch_size", default=None, type=int, help="Batch size. ")
+    parser.add_argument(
+        "--profile",
+        action="store_true",
+        help="Whether to profile the performance using newstest2014 dataset. ")
     args = parser.parse_args()
     return args
 
@@ -77,7 +84,7 @@ def post_process_seq(seq, bos_idx, eos_idx, output_bos=False, output_eos=False):
 
 def do_predict(args):
     place = "gpu"
-    paddle.set_device(place)
+    place = paddle.set_device(place)
 
     # Define data loader
     test_loader, to_tokens = reader.create_infer_loader(args)
@@ -110,21 +117,39 @@ def do_predict(args):
 
     f = open(args.output_file, "w")
     with paddle.no_grad():
+        if args.profile:
+            import time
+            start = time.time()
         for (src_word, ) in test_loader:
             finished_seq = transformer(src_word=src_word)
+            if not args.profile:
+                if args.decoding_strategy == "beam_search":
+                    finished_seq = finished_seq.numpy().transpose([1, 2, 0])
+                elif args.decoding_strategy == "topk_sampling" or args.decoding_strategy == "topp_sampling":
+                    finished_seq = np.expand_dims(
+                        finished_seq.numpy().transpose([1, 0]), axis=1)
+                for ins in finished_seq:
+                    for beam_idx, beam in enumerate(ins):
+                        if beam_idx >= args.n_best:
+                            break
+                        id_list = post_process_seq(beam, args.bos_idx,
+                                                   args.eos_idx)
+                        word_list = to_tokens(id_list)
+                        sequence = " ".join(word_list) + "\n"
+                        f.write(sequence)
+        if args.profile:
             if args.decoding_strategy == "beam_search":
-                finished_seq = finished_seq.numpy().transpose([1, 2, 0])
-            elif args.decoding_strategy == "topk_sampling" or args.decoding_strategy == "topp_sampling":
-                finished_seq = np.expand_dims(
-                    finished_seq.numpy().transpose([1, 0]), axis=1)
-            for ins in finished_seq:
-                for beam_idx, beam in enumerate(ins):
-                    if beam_idx >= args.n_best:
-                        break
-                    id_list = post_process_seq(beam, args.bos_idx, args.eos_idx)
-                    word_list = to_tokens(id_list)
-                    sequence = " ".join(word_list) + "\n"
-                    f.write(sequence)
+                logger.info("Setting info: batch size: {}, beam size: {}. ".
+                            format(args.infer_batch_size, args.beam_size))
+            elif args.decoding_strategy == "topk_sampling":
+                logger.info("Setting info: batch size: {}, topk: {}. ".format(
+                    args.infer_batch_size, args.topk))
+            elif args.decoding_strategy == "topp_sampling":
+                logger.info("Setting info: batch size: {}, topp: {}. ".format(
+                    args.infer_batch_size, args.topp))
+            paddle.fluid.core._cuda_synchronize(place)
+            logger.info("Average time latency is {} ms/batch. ".format((
+                time.time() - start) / len(test_loader) * 1000))
 
 
 if __name__ == "__main__":
@@ -132,13 +157,16 @@ if __name__ == "__main__":
     yaml_file = ARGS.config
     with open(yaml_file, 'rt') as f:
         args = AttrDict(yaml.safe_load(f))
-        pprint(args)
     args.decoding_lib = ARGS.decoding_lib
     args.use_fp16_decoding = ARGS.use_fp16_decoding
     args.decoding_strategy = ARGS.decoding_strategy
     args.beam_size = ARGS.beam_size
     args.topk = ARGS.topk
     args.topp = ARGS.topp
+    args.profile = ARGS.profile
     args.benchmark = False
+    if ARGS.batch_size:
+        args.infer_batch_size = ARGS.batch_size
+    pprint(args)
 
     do_predict(args)