<!-- Demo: https://github.com/PaddlePaddle/PaddleNLP/pull/26 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Others

### PR changes
<!-- One of [ Models | APIs | Docs | Others ] -->
Others

### Description
<!-- Describe what this PR does -->
[FasterTransformer] Update FasterTransformer and add performance info logger. 

The detail performance info will be printed in this PR later. Average latency for 3003 sentences in `en-de` Translation. 

| hyper params | fp32(in ms/batch) | fp16(in ms/batch) | Original(in ms/batch) | speedup(compare with fp32) |
|--|--|--|--|--|
| batch_size: 1, beam_size: 1 | 36.6575 | 41.9541 | 1288.0803@enlei | 35.1382@enlei |
| batch_size: 1, beam_size: 4 | 49.5412 | 46.2929 | 1232.0821 | 24.87 |
| batch_size: 8, beam_size: 1 | 67.2444 | 64.3771 | 1997.2295 | 29.70 |
| batch_size: 8, beam_size: 4 | 75.6515 | 66.5741 | 2135.4212 | 28.23 |
| batch_size: 32, beam_size: 1 | 97.9693 | 84.9347 | 2633.8843 | 26.88 |
| batch_size: 32, beam_size: 4 | 184.9824 | 133.9078 | 2606.9408 | 14.09 |
| batch_size: 64, beam_size: 1 | 152.0229 | 122.5850 | 2881.6411 | 18.96 |
| batch_size: 64, beam_size: 4 | 347.8825 | 218.8199 | 3391.2002 | 9.75 |

**Note**:
* `fp16` means using `fp16` to calculate `decoding` while using `fp32` to calculate `encoder`. Hence, there is a `paddle.cast` casting the output of `encoder` to `fp16`, which may lead to performance drop. 
* The larger `batch_size`/`beam_size`, the less speedup. Since with the increasing number of `batch_size`/`beam_size`, the bottleneck of performance is no longer launching an op/kernel, but the computing in an op/kernel, which cannot be solved by fusion.